✨ migrate pinning store to pinia
