+++ components/rmrk/Create/CreateCollection.vue
@@ -60,7 +60,7 @@ import AuthMixin from '@/utils/mixins/authMixin'
 import MetaTransactionMixin from '@/utils/mixins/metaMixin'
 import RmrkVersionMixin from '@/utils/mixins/rmrkVersionMixin'
 import UseApiMixin from '@/utils/mixins/useApiMixin'
-import { PinningKey, pinFileToIPFS, pinJson } from '@/services/nftStorage'
+import { pinFileToIPFS, pinJson } from '@/services/nftStorage'
 import { notificationTypes, showNotification } from '@/utils/notification'
 import { canSupport } from '@/utils/support'
 import {
@@ -74,6 +74,7 @@ import {
   unSanitizeIpfsUrl,
 } from '@kodadot1/minimark'
 import { Component, Ref, mixins } from 'nuxt-property-decorator'
+import { usePinningStore } from '@/stores/pinning'
 
 type BaseCollectionType = {
   name: string
@@ -135,6 +136,10 @@ export default class CreateCollection extends mixins(
     return Number(this.balance) <= 2
   }
 
+  get pinningStore() {
+    return usePinningStore()
+  }
+
   public constructRmrkMint(metadata: string): CreatedCollection {
     const {
       symbol,
@@ -147,11 +152,7 @@ export default class CreateCollection extends mixins(
 
   public async constructMeta() {
     const { file, name, description } = this.base
-
-    const pinningKey: PinningKey = await this.$store.dispatch(
-      'pinning/fetchPinningKey',
-      this.accountId
-    )
+    const pinningKey = await this.pinningStore.fetchPinningKey(this.accountId)
 
     const imageHash = !file
       ? IPFS_KODADOT_IMAGE_PLACEHOLDER
